Add a deployment step to wake up sleeping serverless DBs
In dev/qa, the RDS instance will scale down to zero capacity units when it's inactive for a bit.

Deployments fail because it takes a couple seconds to wake up when you want to use it --> migrations time out.

This command serves the DB some coffee ahead of the migrations.