Permit components to validate-as-multiple in single mode.
The File component's validated submissionValue() returns data that is structured differently from the JSON submission document.

This causes form.io to be unable to display/edit files when shown after saving, if you're persisting the validated/cleaned data.

I resolve this by adding a new method for validate() to use when determining if we need to validate in single or multi-value mode. File forces itself to always be validated as multiple values, but is still presenting to the outside world as a single-value field when configured that way.