Handle potential special characters in the survey questions
The Formio builder lets you enter something like "Foo bar. Baz!" as the question's value.

When validations, dynamic-forms uses the value to build the path (surveyComponentKey.questionValue), so having the dots makes it think the data is nested more than it actually is in the submission.

This PR fixes that by slugging the question values internally, but it should not impact the output.